Maintenance of imgid
The usage of an `imgid` all over dt code differs significantly, from my current understanding

- an `imgid` must always be a signed integer, by common usage we define as int32_t,
  at many places we use `gint` or `int` which doesn't matter on used 64bit platforms.
- a valid `imgid` is always greater than zero

At some places we used uint32_t which seems to be wrong for me

The tests for a valid imgid should then be (imgid > 0), at some places we tested by `(imgid != 0)` ...

+++ src/common/image_cache.c
@@ -206,7 +206,7 @@ void dt_image_cache_print(dt_image_cache_t *cache)
 dt_image_t *dt_image_cache_get(dt_image_cache_t *cache, const int32_t imgid, char mode)
 {
   if(imgid <= 0) return NULL;
-  dt_cache_entry_t *entry = dt_cache_get(&cache->cache, (uint32_t)imgid, mode);
+  dt_cache_entry_t *entry = dt_cache_get(&cache->cache, imgid, mode);
   ASAN_UNPOISON_MEMORY_REGION(entry->data, sizeof(dt_image_t));
   dt_image_t *img = (dt_image_t *)entry->data;
   img->cache_entry = entry;
@@ -216,7 +216,7 @@ dt_image_t *dt_image_cache_get(dt_image_cache_t *cache, const int32_t imgid, cha
 dt_image_t *dt_image_cache_testget(dt_image_cache_t *cache, const int32_t imgid, char mode)
 {
   if(imgid <= 0) return NULL;
-  dt_cache_entry_t *entry = dt_cache_testget(&cache->cache, (uint32_t)imgid, mode);
+  dt_cache_entry_t *entry = dt_cache_testget(&cache->cache, imgid, mode);
   if(!entry) return 0;
   ASAN_UNPOISON_MEMORY_REGION(entry->data, sizeof(dt_image_t));
   dt_image_t *img = (dt_image_t *)entry->data;
@@ -341,7 +341,9 @@ void dt_image_cache_set_change_timestamp(dt_image_cache_t *cache, const int32_t 
   dt_image_cache_write_release(cache, img, DT_IMAGE_CACHE_SAFE);
 }
 
-void dt_image_cache_set_change_timestamp_from_image(dt_image_cache_t *cache, const int32_t imgid, const int32_t sourceid)
+void dt_image_cache_set_change_timestamp_from_image(dt_image_cache_t *cache,
+                                                    const int32_t imgid,
+                                                    const int32_t sourceid)
 {
   if(imgid <= 0 || sourceid <= 0) return;
